Aligning with working copy

+++ bacnet/demo/epics/main.c
@@ -1446,7 +1446,7 @@ int main(
     BACNET_ADDRESS src; /* address where message came from */
     uint16_t pdu_len = 0;
     unsigned timeout = 100;     /* milliseconds */
-    unsigned max_apdu = 0;
+    uint16_t max_apdu = 0;
     time_t elapsed_seconds = 0;
     time_t last_seconds = 0;
     time_t current_seconds = 0;
