+++ bacnet/demo/handler/s_rd.c
@@ -75,7 +75,7 @@ uint8_t Send_Reinitialize_Device_Request(
 {
     BACNET_ADDRESS dest;
     BACNET_ADDRESS my_address;
-    unsigned max_apdu = 0;
+    uint16_t max_apdu = 0;
     uint8_t invoke_id = 0;
     bool status = false;
     int len = 0;
