+++ bacnet/include/cov.h
@@ -60,7 +60,7 @@ typedef struct BACnet_COV_Subscription_Flags {
 typedef struct {
     BACNET_COV_SUBSCRIPTION_FLAGS flag;
     int8_t dest_index;      // Has to be signed, tested for < 0 in places
-    uint8_t invokeID;   /* for confirmed COV */
+    uint8_t invokeID;       /* for confirmed COV */
     uint32_t subscriberProcessIdentifier;
     uint32_t lifetime;  /* optional */
     BACNET_OBJECT_ID monitoredObjectIdentifier;
@@ -79,38 +79,38 @@ struct BACnet_Subscribe_COV_Data;
 typedef struct BACnet_Subscribe_COV_Data {
     uint32_t subscriberProcessIdentifier;
     BACNET_OBJECT_ID monitoredObjectIdentifier;
-    bool cancellationRequest;   /* true if this is a cancellation request */
+    bool cancellationRequest;			/* true if this is a cancellation request */
     bool issueConfirmedNotifications;   /* optional */
-    uint32_t lifetime;  /* seconds, optional */
+    uint32_t lifetime;					/* seconds, optional 2^32 seconds = 136 years */
     BACNET_PROPERTY_REFERENCE monitoredProperty;
-    bool covIncrementPresent;   /* true if present */
-    float covIncrement; /* optional */
+    bool covIncrementPresent;			/* true if present */
+    float covIncrement;					/* optional */
     BACNET_ERROR_CLASS error_class;
     BACNET_ERROR_CODE error_code;
     struct BACnet_Subscribe_COV_Data *next;
 } BACNET_SUBSCRIBE_COV_DATA;
 
 
-    int ucov_notify_encode_apdu(
-        uint8_t * apdu,
-        unsigned max_apdu_len,
-        BACNET_COV_DATA * data);
+int ucov_notify_encode_apdu(
+    uint8_t * apdu,
+    uint16_t max_apdu_len,
+    BACNET_COV_DATA * data);
 
-    int ucov_notify_decode_apdu(
-        uint8_t * apdu,
-        unsigned apdu_len,
-        BACNET_COV_DATA * data);
+int ucov_notify_decode_apdu(
+    uint8_t * apdu,
+    unsigned apdu_len,
+    BACNET_COV_DATA * data);
 
-    int ucov_notify_send(
-        uint8_t * apdu,
-        unsigned max_apdu_len,
-        BACNET_COV_DATA * data);
+int ucov_notify_send(
+    uint8_t * apdu,
+    uint16_t max_apdu_len,
+    BACNET_COV_DATA * data);
 
-    int ccov_notify_encode_apdu(
-        uint8_t * apdu,
-        unsigned max_apdu_len,
-        uint8_t invoke_id,
-        BACNET_COV_DATA * data);
+int ccov_notify_encode_apdu(
+    uint8_t * apdu,
+    uint16_t max_apdu_len,
+    uint8_t invoke_id,
+    BACNET_COV_DATA * data);
 
 int ccov_notify_decode_apdu(
     uint8_t * apdu,
@@ -129,22 +129,22 @@ int cov_subscribe_property_decode_service_request(
     unsigned apdu_len,
     BACNET_SUBSCRIBE_COV_DATA * data);
 
-    int cov_subscribe_property_encode_apdu(
-        uint8_t * apdu,
-        unsigned max_apdu_len,
-        uint8_t invoke_id,
-        BACNET_SUBSCRIBE_COV_DATA * data);
+int cov_subscribe_property_encode_apdu(
+    uint8_t * apdu,
+    uint16_t max_apdu_len,
+    uint8_t invoke_id,
+    BACNET_SUBSCRIBE_COV_DATA * data);
 
 int cov_subscribe_decode_service_request(
     uint8_t * apdu,
     unsigned apdu_len,
     BACNET_SUBSCRIBE_COV_DATA * data);
 
-    int cov_subscribe_encode_apdu(
-        uint8_t * apdu,
-        unsigned max_apdu_len,
-        uint8_t invoke_id,
-        BACNET_SUBSCRIBE_COV_DATA * data);
+int cov_subscribe_encode_apdu(
+    uint8_t * apdu,
+    uint16_t max_apdu_len,
+    uint8_t invoke_id,
+    BACNET_SUBSCRIBE_COV_DATA * data);
 
 void cov_data_value_list_link(
     BACNET_COV_DATA *data,
